+++ src/ai_bom_generator/exporters/__init__.py
@@ -1,3 +1,4 @@
 from ai_bom_generator.exporters.cyclonedx_json.mapper import export_cyclonedx_json
+from ai_bom_generator.exporters.spdx_ai.mapper import export_spdx_ai
 
-__all__ = ["export_cyclonedx_json"]
+__all__ = ["export_cyclonedx_json", "export_spdx_ai"]